add  remote

+++ database/migrations/2022_01_24_012246_create_trainers_table.php
@@ -16,7 +16,7 @@ class CreateTrainersTable extends Migration
         Schema::create('trainers', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->long('harga');
+            $table->float('harga');
             $table->string('classname');
             $table->timestamps();
         });
